requirements.txt
telethon
pyYAML